project1/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Jinja2

python-dotenv

azure-cognitiveservices-vision-computervision
msrest

google-generativeai  
diffusers             
transformers
accelerate
safetensors
Pillow

requests

torch

azure-cognitiveservices-vision-customvision
